feat: 말풍선에 waveform 이식을 위한 매개변수 및 변수 추가, 파일 분리
